Separate MembersTextFormField from TitleTextFormField
